Think or Swim strategies Auto B, sHI, S,SL
This strategy has a better chance of opening a position with profit. Macd study Value line must pass the zero line which build support for a buying signal. Setting up for the closing positions are three close orders Sell Hi, sell, stop loss. The sell high will close once the study for the bollinger band price crosses the upper band which signals high volatility. Which takes advantage of Theta. The Sell order closes the positions if MACD value line crosses the average line, this will close the position and cut the chance of a loss. Last is the Stop loss, this will take a minimize the loss.